Fixed the pointer issue
Could someone please run AttackerDriver.cpp to make sure everything is alright? I've become rather blind to my own code by this point.
You can comment out lines 110-175 of the driver to just get right to the attacking loop which is where the initial issue of the countries not transfering properly was.